+++ .projenrc.js
@@ -49,18 +49,18 @@ const project = new JsiiProject({
 
   // makes it very hard to iterate with jest --watch
   jestOptions: {
-    coverageText: false
+    coverageText: false,
   },
 
   publishToMaven: {
     javaPackage: 'org.projen',
     mavenGroupId: 'com.github.eladb',
-    mavenArtifactId: 'projen'
+    mavenArtifactId: 'projen',
   },
 
   publishToPypi: {
     distName: 'projen',
-    module: 'projen'
+    module: 'projen',
   },
 });